refactor(cli): split commands.rs (4.2k lines) into per-group modules (#190)
Pure move: aifw-cli/src/commands/{common,rules,system,nat,shaping,vpn,
geoip,config,routes,dns,users,interfaces,dhcp,updates,reverse_proxy,
multiwan,cluster,ids,logging}.rs with a thin mod.rs re-exporting every
public fn, so main.rs is untouched. Shared parsers/engine constructors
live in common.rs (pub(super)); the two cross-group helpers
(cluster's API base/token, shaping's engine ctor) are pub(super) and
imported explicitly. Unused imports pruned per file; no behaviour change.
