+++ aifw-cli/src/commands/common.rs
@@ -1,0 +1,54 @@
+//! Shared parsers and engine constructors used by several command groups.
+
+use aifw_common::{Action, Direction, PortRange};
+use aifw_core::{Database, NatEngine, RuleEngine};
+use std::path::Path;
+use std::sync::Arc;
+
+pub(super) fn parse_port(s: &str) -> anyhow::Result<PortRange> {
+    if let Some((start, end)) = s.split_once(':') {
+        Ok(PortRange {
+            start: start.parse()?,
+            end: end.parse()?,
+        })
+    } else {
+        let port: u16 = s.parse()?;
+        Ok(PortRange {
+            start: port,
+            end: port,
+        })
+    }
+}
+
+pub(super) fn parse_action(s: &str) -> anyhow::Result<Action> {
+    match s {
+        "pass" => Ok(Action::Pass),
+        "block" => Ok(Action::Block),
+        "block-drop" => Ok(Action::BlockDrop),
+        "block-return" => Ok(Action::BlockReturn),
+        _ => anyhow::bail!("unknown action: {s} (use pass, block, block-drop, block-return)"),
+    }
+}
+
+pub(super) fn parse_direction(s: &str) -> anyhow::Result<Direction> {
+    match s {
+        "in" => Ok(Direction::In),
+        "out" => Ok(Direction::Out),
+        "any" => Ok(Direction::Any),
+        _ => anyhow::bail!("unknown direction: {s} (use in, out, any)"),
+    }
+}
+
+pub(super) async fn create_engine(db_path: &Path) -> anyhow::Result<RuleEngine> {
+    let db = Database::new(db_path).await?;
+    let pf = Arc::from(aifw_pf::create_backend());
+    Ok(RuleEngine::new(db.pool().clone(), pf))
+}
+
+pub(super) async fn create_nat_engine(db_path: &Path) -> anyhow::Result<NatEngine> {
+    let db = Database::new(db_path).await?;
+    let pf = Arc::from(aifw_pf::create_backend());
+    // "aifw-nat" — must match the API/daemon anchor; NAT loads replace every
+    // rule class in their anchor since #531.
+    Ok(NatEngine::new(db.pool().clone(), pf).with_anchor("aifw-nat".to_string()))
+}
